rename assignseat to appoverview
