+++ frontend/src/components/AdminApplicationOverview/constants.ts
