+++ frontend/src/components/AdminApplicationOverview/reducer.ts
