+++ frontend/src/components/AdminApplicationOverview/strings.ts
