+++ frontend/src/components/Route/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
+import AdminApplicationOverview from '../AdminApplicationOverview';
 import AdminRoom from '../AdminRoom';
 import Application from '../Application';
-import AssignSeat from '../AssignSeat';
 import CreateSeason from '../CreateSeason';
 import { Header } from '../Header';
 import Home from '../Home';
@@ -22,7 +22,7 @@ export const Routing = () => {
           <Route exact={true} path="/admin/rooms/create-room" component={AdminRoom} />
           <Route path="/admin/create-season" component={CreateSeason} />
           <Route path="/admin/rooms/update-room" component={AdminRoom} />
-          <Route path="/admin/applications" component={AssignSeat} />
+          <Route path="/admin/applications" component={AdminApplicationOverview} />
         </Switch>
       </div>
     </>
